backend/requirements.txt
Django==2.2.7
pytz==2019.3
sqlparse==0.3.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
